+++ src/main/java/org/smoothbuild/cli/command/BuildCommand.java
@@ -43,9 +43,9 @@ public class BuildCommand extends ExclusiveCommand {
           least warning level.
           
           Available task matchers:
-            all              - all tasks
-            default          - shortcut for 'info|(user&(call|value|field))'
-            none             - no tasks
+            a, all           - all tasks
+            d, default       - shortcut for 'info|(user&(call|value|field))'
+            n, none          - no tasks
           
             f, fatal         - contains a log with fatal level
             e, error         - contains a log with at least error level